Use remote owner if only search result
Since it is highly likely that the username from the remote URL matches
a specific GitHub username that the user would like to view, show if
it's the only result. Subcommand renamed to "by-remote" since, from the
user's perspective, they may not need to select from a search (also just
makes the subcommand less work to type).